requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
python-docx>=1.0.1
PyPDF2>=3.0.0
sentence-transformers
numpy<2.0.0
pandas
python-dotenv
watchdog
requests>=2.31.0
aiofiles>=23.2.1
python-multipart>=0.0.6
aiofiles>=24.0.0
